update delete to only authorized user

+++ part4/blog/controllers/blogs.js
@@ -39,6 +39,20 @@ blogsRouter.post("/", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response) => {
 	const id = request.params.id;
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
+	const user = await User.findById(decodedToken.id);
+
+	user.blogs = user.blogs.filter(
+		(blogId) => blogId.toString() !== id.toString()
+	);
+
+	await user.save();
+
 	await Blog.findByIdAndDelete(id);
 	return response.status(204).end();
 });
